+++ saltstack/database/python_classes/db_crud_device.py
@@ -9,7 +9,7 @@ db = client.test
 
 
 def main():
-    while (1):
+    while 1:
         selection = raw_input('\nSelect 1 to insert, 2 to update, 3 to read, 4 to delete, "CTRL +C" to end\n')
 
         if selection == '1':
@@ -22,12 +22,16 @@ def main():
             print 'delete'
             delete()
         else:
-            print '\n INVALID SELECTION \n'
+            print '\nINVALID SELECTION\n'
 
 def insert():
+    host_name = ''
+    ip_address = ''
     try:
-        host_name = raw_input('Enter Device Host name :')
-        ip_address = raw_input('Enter IP Address :')
+        while not host_name:
+            host_name = raw_input('Enter Device Host name *required:')
+        while not ip_address:
+            ip_address = raw_input('Enter IP Address *required:')
         mac_address = raw_input('Enter MAC Address :')
         dev_class = raw_input('Enter Device Class "Router", "Switch", "Device" :')
         role = raw_input('Enter Device Role "Core", "Distribution", "Access", "External" :')
@@ -35,17 +39,27 @@ def insert():
         port_ip = raw_input('Enter Interface IP Address :')
         port_neighbor = raw_input('Enter Neighbor :')
 
-        db.network.insert_one(
-            {
-              "host_name": host_name,
-              "ip_address": ip_address,
-              "MAC_address": mac_address,
-              "Class": dev_class,
-              "Role": role,
-              "connections": [
-                { "interface": port_name, "ip": port_ip, "neighbor": port_neighbor}
-              ]
-            })
+        while 1:
+            selection = raw_input("\nInsert this into Database?\nHost Name:" + host_name + '[y] or [n]: ')
+            if selection == 'y':
+                db.network.insert_one(
+                    {
+                        "host_name": host_name,
+                        "ip_address": ip_address,
+                        "MAC_address": mac_address,
+                        "Class": dev_class,
+                        "Role": role,
+                        "connections": [
+                            {"interface": port_name, "ip": port_ip, "neighbor": port_neighbor}
+                        ]
+                    })
+                print '\nInserted data successfully\n'
+                break
+            elif selection == 'n':
+                print '\nInsertion Cancelled\n'
+                break
+            else:
+                print '\nINVALID SELECTION\n'
         print '\nInserted data successfully\n'
 
     except Exception, e:
@@ -94,8 +108,18 @@ def read():
 def delete():
     try:
         del_dev = raw_input('\nEnter Device Host Name to delete\n')
-        db.network.delete_many({"host_name": del_dev})
-        print '\nDeletion successful\n'
+        while 1:
+            selection = raw_input('\nDelete every Device with Hostname: ' + del_dev + ", [y] or [n]:")
+
+            if selection == 'y':
+                db.network.delete_many({"host_name": del_dev})
+                print '\nDeletion successful\n'
+                break
+            elif selection == 'n':
+                print '\nDeletion cancelled\n'
+                break
+            else:
+                print '\nINVALID SELECTION\n'
     except Exception, e:
         print str(e)
